Latest work

+++ WebApplication.Core/Users/Commands/DeleteUserCommand.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using AutoMapper;
 using FluentValidation;
+using FluentValidation.Results;
 using MediatR;
 using WebApplication.Core.Common.Exceptions;
 using WebApplication.Core.Users.Common.Models;
@@ -37,6 +38,10 @@ namespace WebApplication.Core.Users.Commands
             /// <inheritdoc />
             public async Task<UserDto> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
             {
+                Validator validator = new Validator();
+                ValidationResult validationResult = validator.Validate(request);
+                if (!validationResult.IsValid) throw new ValidationException(validationResult.Errors);
+
                 User? deletedUser = await _userService.DeleteAsync(request.Id, cancellationToken);
 
                 if (deletedUser is default(User)) throw new NotFoundException($"The user '{request.Id}' could not be found.");
